dataseeder/pyproject.toml
[project]
name = "bellhisfirm-dataseeder"
version = "1.0.0"
description = "Data seeder for BellHisFirm database - generates realistic Belgian company and person data"
authors = [
    {name = "BellHisFirm Project"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "faker>=22.0.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
seed = "dataseeder.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataseeder"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]
